Extract GitHub review signal module
